+++ changelog.md
@@ -4,7 +4,10 @@ October 29
 - Researched some new methods/workflow
 - Not need python flask for now
 - Will use React + Vite workflow
-- install node.js
+- install node.js react vite
+- set up dev server
+- work on html css to this new workflow
+- have it functional
 
 
 October 28